plugins, command line parsing, url parsing, initial test, etc.
This is still work in progress, but the skeleton3/desktop now has a
command line parser :) It also has a first simple test to do some
testing on topological sort algorithm. Filename to load must now be
specified as part of the url as ...&file=/tmp/tree.png for example.

Still need to add command line parsing to server.