CAS-9971 fix a bug that, Open green.fits then open another file, crash.
The root cause is, it is possible that the caller will pass incompatible displayAxisTypes into DataSource::_setDisplayAxes for any astro file, so the indices of axes become wired [0,0]. In the end, it will crash on CARTA_ASSERT( usedCount == 0 ); in CCImage.h.

Solution:
Add a check in DataSource::_setDisplayAxes, do not use incompatible displayAxisTypes.